rename operation definition helper and split out creator and finders to seperate modules
