+++ nhais-adaptor/adaptor/outgoing/fhir_helpers/tests/test_fhir_creators.py
@@ -1,15 +1,14 @@
 import unittest
 
 from fhirclient.models.patient import Patient
-from testfixtures import compare
 from fhirclient.models.operationdefinition import OperationDefinitionParameter, OperationDefinition
 from fhirclient.models.practitioner import Practitioner
 
-from adaptor.outgoing.fhir_helpers.operation_definition import OperationDefinitionHelper as odh
-from adaptor.outgoing.fhir_helpers.constants import ParameterName, ResourceType
+import adaptor.outgoing.fhir_helpers.fhir_creators as creators
+from adaptor.outgoing.fhir_helpers.constants import ResourceType
 
 
-class OperationDefinitionHelperTest(unittest.TestCase):
+class TestFhirCreators(unittest.TestCase):
     """
     Tests the helper functions for the fhir operation definition
     """
@@ -18,7 +17,7 @@ class OperationDefinitionHelperTest(unittest.TestCase):
         """
         Test the helper function to create an operation parameter with a binding value
         """
-        op_param = odh.create_parameter_with_binding(name="SOME_NAME", value="SOME_VALUE")
+        op_param = creators.create_parameter_with_binding(name="SOME_NAME", value="SOME_VALUE")
 
         self.assertIsInstance(op_param, OperationDefinitionParameter)
         self.assertEqual(op_param.name, "SOME_NAME")
@@ -28,7 +27,7 @@ class OperationDefinitionHelperTest(unittest.TestCase):
         """
         Test the helper function to create an operation parameter with a resource reference
         """
-        op_param = odh.create_parameter_with_resource_ref(name="SOME_NAME", resource_type=ResourceType.PRACTITIONER,
+        op_param = creators.create_parameter_with_resource_ref(name="SOME_NAME", resource_type=ResourceType.PRACTITIONER,
                                                           reference="practitioner-1")
 
         self.assertIsInstance(op_param, OperationDefinitionParameter)
@@ -42,7 +41,7 @@ class OperationDefinitionHelperTest(unittest.TestCase):
         """
 
         with self.subTest("with no contained resources or parameters"):
-            op_def = odh.create_operation_definition(name="name.of.operation", code="code.of.operation",
+            op_def = creators.create_operation_definition(name="name.of.operation", code="code.of.operation",
                                                      date_time="2019-04-23 09:00:04.159338", contained=[],
                                                      parameters=[])
 
@@ -54,9 +53,9 @@ class OperationDefinitionHelperTest(unittest.TestCase):
             self.assertEqual(op_def.parameter, None)
 
         with self.subTest("with one parameter"):
-            some_param = odh.create_parameter_with_binding(name="SOME_NAME", value="SOME_VALUE")
+            some_param = creators.create_parameter_with_binding(name="SOME_NAME", value="SOME_VALUE")
 
-            op_def = odh.create_operation_definition(name="name.of.operation", code="code.of.operation",
+            op_def = creators.create_operation_definition(name="name.of.operation", code="code.of.operation",
                                                      date_time="2019-04-23 09:00:04.159338", contained=[],
                                                      parameters=[some_param])
 
@@ -64,10 +63,10 @@ class OperationDefinitionHelperTest(unittest.TestCase):
             self.assertIsInstance(op_def.parameter[0], OperationDefinitionParameter)
 
         with self.subTest("with one contained resource"):
-            some_resource = odh.create_practitioner_resource(resource_id="some_id", national_identifier="some_nat_id",
+            some_resource = creators.create_practitioner_resource(resource_id="some_id", national_identifier="some_nat_id",
                                                              local_identifier="some_local_id")
 
-            op_def = odh.create_operation_definition(name="name.of.operation", code="code.of.operation",
+            op_def = creators.create_operation_definition(name="name.of.operation", code="code.of.operation",
                                                      date_time="2019-04-23 09:00:04.159338", contained=[some_resource],
                                                      parameters=[])
 
@@ -75,7 +74,7 @@ class OperationDefinitionHelperTest(unittest.TestCase):
             self.assertIsInstance(op_def.contained[0], Practitioner)
 
     def test_create_practitioner(self):
-        practitioner = odh.create_practitioner_resource(resource_id="practitioner-1", national_identifier="1234567",
+        practitioner = creators.create_practitioner_resource(resource_id="practitioner-1", national_identifier="1234567",
                                                         local_identifier="111")
 
         self.assertIsInstance(practitioner, Practitioner)
@@ -84,7 +83,7 @@ class OperationDefinitionHelperTest(unittest.TestCase):
         self.assertEqual(practitioner.identifier[1].value, "111")
 
     def test_create_patient(self):
-        patient = odh.create_patient_resource(resource_id="patient-1", nhs_number="NHSNO11111",
+        patient = creators.create_patient_resource(resource_id="patient-1", nhs_number="NHSNO11111",
                                               title="Mr", first_name="Peter", last_name="Parker",
                                               gender="male", date_of_birth="2019-04-23", place_of_birth="Spidey Town",
                                               address_line_1="1 Spidey Way", city="Spidey Town", postcode="SP1 1AA")
@@ -102,25 +101,4 @@ class OperationDefinitionHelperTest(unittest.TestCase):
         self.assertEqual(patient.address[0].city, "Spidey Town")
         self.assertEqual(patient.address[0].postalCode, "SP1 1AA")
 
-    def test_get_parameter_value(self):
-        op_param_transaction_number = odh.create_parameter_with_binding(name="some_param_name", value="17")
 
-        op = odh.create_operation_definition(name="some.name", code="some.code", date_time="2019-04-23 09:00:04.159338",
-                                             contained=[], parameters=[op_param_transaction_number])
-
-        self.assertEqual(odh.get_parameter_value(op, "some_param_name"), "17")
-
-    def test_find_resource(self):
-        with self.subTest("When practitioner details are found in the operation definition"):
-            practitioner = odh.create_practitioner_resource(resource_id="practitioner-1", national_identifier="1234567",
-                                                            local_identifier="111")
-            op_param_practitioner = odh.create_parameter_with_resource_ref(name=ParameterName.REGISTER_PRACTITIONER,
-                                                                           resource_type=ResourceType.PRACTITIONER,
-                                                                           reference="practitioner-1")
-            op_def = odh.create_operation_definition(name="name.of.operation", code="code.of.operation",
-                                                     date_time="2019-04-23 09:00:04.159338", contained=[practitioner],
-                                                     parameters=[op_param_practitioner])
-
-            found_practitioner = odh.find_resource(fhir_operation=op_def, resource_type=ResourceType.PRACTITIONER)
-
-            compare(found_practitioner, practitioner)
