+++ nhais-adaptor/adaptor/outgoing/tests/test_interchange_adaptor.py
@@ -2,7 +2,7 @@ import unittest
 import adaptor.outgoing.fhir_helpers.tests.fixtures as fixtures
 from testfixtures import compare
 from adaptor.outgoing.interchange_adaptor import InterchangeAdaptor
-from adaptor.outgoing.fhir_helpers.operation_definition import OperationDefinitionHelper as odh
+import adaptor.outgoing.fhir_helpers.fhir_creators as creators
 from adaptor.outgoing.fhir_helpers.constants import ParameterName, ResourceType, OperationName
 
 
@@ -51,23 +51,23 @@ class InterchangeAdaptorTest(unittest.TestCase):
                                             "UNT+18+000001'"
                                             "UNZ+1+000001'")
 
-            op_param_interchange_sequence = odh.create_parameter_with_binding(name=ParameterName.INTERCHANGE_SEQ_NO,
+            op_param_interchange_sequence = creators.create_parameter_with_binding(name=ParameterName.INTERCHANGE_SEQ_NO,
                                                                               value="000001")
-            op_param_sender_cypher = odh.create_parameter_with_binding(name=ParameterName.SENDER_CYPHER, value="TES5")
-            op_param_message_sequence = odh.create_parameter_with_binding(name=ParameterName.MESSAGE_SEQ_NO,
+            op_param_sender_cypher = creators.create_parameter_with_binding(name=ParameterName.SENDER_CYPHER, value="TES5")
+            op_param_message_sequence = creators.create_parameter_with_binding(name=ParameterName.MESSAGE_SEQ_NO,
                                                                           value="000001")
-            op_param_nhais_cypher = odh.create_parameter_with_binding(name=ParameterName.NHAIS_CYPHER, value="XX1")
-            op_param_transaction_number = odh.create_parameter_with_binding(name=ParameterName.TRANSACTION_NO,
+            op_param_nhais_cypher = creators.create_parameter_with_binding(name=ParameterName.NHAIS_CYPHER, value="XX1")
+            op_param_transaction_number = creators.create_parameter_with_binding(name=ParameterName.TRANSACTION_NO,
                                                                             value="17")
             practitioner = fixtures.create_simple_practitioner()
             patient = fixtures.create_simple_patient()
-            op_param_practitioner = odh.create_parameter_with_resource_ref(name=ParameterName.REGISTER_PRACTITIONER,
+            op_param_practitioner = creators.create_parameter_with_resource_ref(name=ParameterName.REGISTER_PRACTITIONER,
                                                                            resource_type=ResourceType.PRACTITIONER,
                                                                            reference="practitioner-1")
-            op_param_patient = odh.create_parameter_with_resource_ref(name=ParameterName.REGISTER_PATIENT,
+            op_param_patient = creators.create_parameter_with_resource_ref(name=ParameterName.REGISTER_PATIENT,
                                                                       resource_type=ResourceType.PATIENT,
                                                                       reference="patient-1")
-            op_def = odh.create_operation_definition(name=OperationName.REGISTER_BIRTH, code="gpc.registerpatient",
+            op_def = creators.create_operation_definition(name=OperationName.REGISTER_BIRTH, code="gpc.registerpatient",
                                                      date_time="2019-04-23 09:00:04.159338",
                                                      contained=[practitioner, patient],
                                                      parameters=[op_param_interchange_sequence,
